// src/discrete-layer/components/layer-details/raster/feature-type.enum.ts
export enum FeatureType {
  PP_PERIMETER = 'PP_PERIMETER',
  PP_PERIMETER_MARKER = 'PP_PERIMETER_MARKER',
  SOURCE_EXTENT = 'SOURCE_EXTENT',
  SOURCE_EXTENT_MARKER = 'SOURCE_EXTENT_MARKER',
  EXISTING_PP = 'EXISTING_PP',
  EXISTING_PP_FOOTPRINT_MARKER = 'EXISTING_PP_FOOTPRINT_MARKER',
  LOW_RESOLUTION_PP = 'LOW_RESOLUTION_PP',
  BACKUP_PP = 'BACKUP_PP',
  CHANGED_AREA_OVERLAPPED_PP = 'CHANGED_AREA_OVERLAPPED_PP',
  CHANGED_AREA_ADDED_PP = 'CHANGED_AREA_ADDED_PP',
}